refactor: split runtime aggregate evaluation
